♻️ Adopt tests to new config location
